requirements.txt
# Core dependencies
pyyaml>=6.0
tldextract>=3.4.0
html2text>=2020.1.16
langdetect>=1.0.9

# crawl4ai integration
crawl4ai>=0.1.0

# Async support
aiohttp>=3.8.4
aiofiles>=23.1.0

# CLI support
argparse>=1.4.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
aioresponses>=0.7.4

# URL validation and processing
validators>=0.20.0
beautifulsoup4>=4.12.0